requirements.txt
aiogram
psycopg2
python-dotenv
spotipy